+++ src/session.rs
@@ -8,7 +8,7 @@ fn new_token_string(len: usize) -> String {
     alpha.sample_string(&mut rand::rng(), len)
 }
 
-pub fn new_session(user_id: i32) -> NewSession {
+pub fn new_session(user_id: i64) -> NewSession {
     let expires = Utc::now().naive_utc() + chrono::Duration::hours(1);
     NewSession {
         token: new_token_string(32),
